+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 # Version 19
 
 ### [Current changes]
+- **[Add]** Action has been added to know element size and location
+- **[Add]** Drag and Drop by offset has been added
 - **[Add]** Custom error msg feature
 - **[Change]** Support for the new Authentication System 2.0 has been added.
 - **[Change]** Improved reconnect time of node by 10x-30x once it detects a live server.